full language support for the backend, access blocking, css fixes

+++ templates/files.delete.php
@@ -1,5 +1,7 @@
 <?php
 
+if(!defined('KIRBY')) die('Direct access is not allowed');
+
 if($panel->action != 'delete-file') return;
 
 $action = action::deleteFile();
@@ -15,7 +17,7 @@ if(!$file) go(showurl('files'));
     <?php growl($action) ?>
 
     <fieldset>
-      <h3>Delete this file</h3>
+      <h3><?php echo l::get('files.delete.title') ?></h3>
       
       <div class="field">
         <label><?php echo html($file->filename()) ?></label>
@@ -23,8 +25,8 @@ if(!$file) go(showurl('files'));
 
       <div class="buttons">
         <input type="hidden" name="filename" value="<?php echo $file->filename() ?>" />
-        <input type="submit" name="delete-file" value="Delete" />
-        <input class="cancel" type="submit" name="cancel-delete-file" value="Cancel" />
+        <input type="submit" name="delete-file" value="<?php echo l::get('files.delete.button') ?>" />
+        <input class="cancel" type="submit" name="cancel-delete-file" value="<?php echo l::get('cancel') ?>" />
       </div>
                   
     </fieldset>
