+++ templates/pages.delete.php
@@ -1,5 +1,7 @@
 <?php
 
+if(!defined('KIRBY')) die('Direct access is not allowed');
+
 if($panel->action != 'delete-page') return;
 
 $action = action::deleteContent();
@@ -15,17 +17,16 @@ if(!$data) go(showurl('pages'));
     <?php growl($action) ?>
 
     <fieldset>
-      <h3>Delete this page</h3>
+      <h3><?php echo l::get('pages.delete.title') ?></h3>
       
       <div class="field">
         <label><?php echo html($data->title()) ?></label>
       </div>
 
       <div class="buttons">
-        <input type="submit" name="delete-page" value="Delete" />
-        <input class="cancel" type="submit" name="cancel-delete-page" value="Cancel" />
+        <input type="submit" name="delete-page" value="<?php echo l::get('pages.delete.button') ?>" />
+        <input class="cancel" type="submit" name="cancel-delete-page" value="<?php echo l::get('cancel') ?>" />
       </div>
-                  
     </fieldset>  
 
   </form>
